Extract thumbnails
The second of the video to use for the thumbnail should be entered
manually to the scene data. The extract_thumbnails script then
saves a single frame of the video as an image for each scene. The
files are named using the second of the video they are taken from.